Fix liking and creating new blog
There was a problem with expanding a blog. After a like or after the
creation of a blog it did not say who added the blog. For likes this was
kind of fixed by automatically refresh the page, but this was far from
ideal since this is noticable and it still takes about 2 seconds after
that. For adding a new blog, the page had to be refreshed manually. If a
newly created blog was liked without refreshing the page, the app would
stop working. The only way to fix this was to remove the blog manually
from the database.

This is not necessary any more, thanks to the getAllUsers method in the
blogService. Adding a like or creating a new blog will now work
properly and refreshing the page is no longer necessary.